Improved calorie calculation and fixed a few bugs.

+++ app/src/main/java/com/sport/smartwatch/Profile.kt
@@ -3,6 +3,6 @@ package com.sport.smartwatch
 open class Profile(
     internal var name: String = "",
     internal var weight: Float = 0.0F,
-    internal var age: Float = 0.0F,
+    internal var age: Int = 0,
     internal var gender: String = "") {
 }
